+++ src/main/java/de/alpharogroup/db/entity/auditable/Auditable.java
@@ -24,18 +24,19 @@
  */
 package de.alpharogroup.db.entity.auditable;
 
-import de.alpharogroup.db.entity.creatable.Creatable;
+import de.alpharogroup.db.entity.creatable.ByCreatable;
+import de.alpharogroup.db.entity.modifiable.ByLastModified;
 import de.alpharogroup.db.entity.modifiable.LastModified;
 
 /**
- * The interface {@link Auditable} is a combination of the interfaces {@link Creatable} and
- * {@link LastModified} and can be implemented by classes with auditing purposes.
+ * The interface {@link Auditable} is a combination of the interfaces {@link ByCreatable} and
+ * {@link ByLastModified} and can be implemented by classes with auditing purposes.
  *
  * @param <T>
  *            the generic type of time measurement
  * @param <U>
  *            the generic type of the user or account
  */
-public interface Auditable<T, U> extends Creatable<T, U>, LastModified<T, U>
+public interface Auditable<T, U> extends ByCreatable<T, U>, ByLastModified<T, U>
 {
 }