+++ src/main/java/de/alpharogroup/db/entity/treeable/IdentifiableTreeable.java
@@ -24,9 +24,9 @@
  */
 package de.alpharogroup.db.entity.treeable;
 
-import java.io.Serializable;
-
 import de.alpharogroup.db.entity.Identifiable;
+
+import java.io.Serializable;
 
 /**
  * The interface {@link IdentifiableTreeable} is a combination of the interfaces