+++ src/main/java/de/alpharogroup/db/entity/validatable/IdentifiableValidatable.java
@@ -24,9 +24,9 @@
  */
 package de.alpharogroup.db.entity.validatable;
 
-import java.io.Serializable;
-
 import de.alpharogroup.db.entity.Identifiable;
+
+import java.io.Serializable;
 
 /**
  * The interface {@link IdentifiableValidatable} is a combination of the interfaces